requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
streamlit
chromadb
python-dotenv
playwright
beautifulsoup4
helium
unstructured
pyhton-magic
python-pptx